US home values by region

A dataset containing seasonally adjusted, monthly estimated typical home values within the middle 30th percentiles (35th to 65th) for regions across the United States

Usage

ushs

Format

Data frame with 255,970 observations of 7 variables:

RegionID Integer, numeric region identifier

Size_Rank Integer, rank of region size

RegionName Character, name of region

RegionType Character, one of c("country", "msa")

StateName Character, name of state in which region falls, if applicable

RecordDate Date, time for which home values were estimated

TypicalHomeValue Numeric, estimated mean value, in US dollars, of typical homes within region and time period

Source

Zillow Home Value Index https://files.zillowstatic.com/research/public_csvs/zhvi/Metro_zhvi_uc_sfrcondo_tier_0.33_0.67_month.csv?t=1700060680